Move sanitizer corpus and mutation parsing checks to real browsers
