+++ docs/test-specs/browser-tests.md
@@ -68,3 +68,13 @@ The eight attack categories retain two sanitizer calls each, recursive inspectio
 This step changes no dependency versions, application behavior, output policy, database or browser assets. Rollback restores the old test file and removes its mapped browser file together. It does not remove jsdom yet.
 
 Against browser-foundation parent `eb01da5b`, the full local browser suite passes 23 cases in Chromium and WebKit on both Node 22.23.2 and 24.20.0. The Node suite passes 1,989 cases with the same three existing pending tests (2,004 minus the 15 migrated cases); client-core passes 283 and dependency compatibility passes 298 (313 minus 15). A clean install/production build and changed-file lint pass. Firefox validation is provided by the required Linux CI jobs because its downloaded binary cannot launch a profile on this Mac. This is a coverage migration, with no additional package or production-memory claim.
+
+## Sanitizer corpus and production reparsing migration
+
+`tests/browser/sanitizer-differential.test.js` replaces the 300 cases in the former top-level differential suite and the one browser-parsing case in `tests/security.test.js`. All 30 corpus objects are unchanged, and all 300 original case titles remain, beneath a new browser-suite parent. The mapping is 120 sanitizer-output safety cases (30 inputs × four sanitizers), 30 divergence records, 120 cases using the existing EJS escaping model, 30 raw `textContent` cases, plus the production case with 12 hostile payloads and three parse/serialize rounds each. Server sanitizer and storage contracts remain in the Node security suite.
+
+Production, xss and strict sanitize-html outputs are still computed with their actual Node modules; DOMPurify runs its installed browser distribution. Each parser case gets a fresh isolated context. `DOMParser` parses an inert full HTML document, preserving the old full-document parser check without executing attack markup in the fixture page. This checks parsed structure, not script execution, layout or a complete application rendering flow. The EJS helper remains the same escaping model as before; actual output-sink and rendering integration tests are separate migration work. The diagnostic divergence report tolerates incomplete matrices during focused or failed runs without suppressing case failures.
+
+On the initial candidate, all 301 cases pass in Chromium and WebKit on Node 22.23.2. The 30 DOMPurify outputs match the previous jsdom results exactly in both engines, with zero attempted network requests during the isolated output comparison. A temporary production-sanitizer bypass fails 11 corpus assertions for script tags, handlers and unsafe URLs; bypassing the sanitizer in the repeated-parsing case also fails. The existing xss less-than-note truncation remains visible in the divergence record; xss is a test comparator, not a proposed runtime replacement.
+
+The remaining main suite passes 1,688 cases with three existing pending tests (1,989 minus 301); no tests were disabled. NYC still reports the same covered line, branch and function counts, with one fewer covered statement in that Node job as these cases move to a separate browser job. Browser coverage is not included in NYC. No application code, dependency version, stored-data policy or browser assets change. Current-parent Chromium/Firefox/WebKit CI on both Node floors, backend checks and native Docker validation remain merge gates. Rollback restores both former test locations and removes the mapped browser file together.
